+++ Kernel/Modules/AgentTicketZoom.pm
@@ -54,7 +54,7 @@ sub new {
             && !defined $Self->{ZoomTimeline}
             )
         {
-            $Self->{ZoomExpand} = $ConfigObject->Get('Ticket::Frontend::ZoomExpand');
+            $Self->{ZoomExpand} = $ConfigObject->Get('Ticket::Frontend::AgentZoomExpand');
             if ( $UserPreferences{UserLastUsedZoomViewType} ) {
                 if ( $UserPreferences{UserLastUsedZoomViewType} eq 'Expand' ) {
                     $Self->{ZoomExpand} = 1;
@@ -89,7 +89,7 @@ sub new {
                     $Self->{ZoomTimeline} = 1;
                 }
                 else {
-                    $LastUsedZoomViewType = $ConfigObject->Get('Ticket::Frontend::ZoomExpand')
+                    $LastUsedZoomViewType = $ConfigObject->Get('Ticket::Frontend::AgentZoomExpand')
                         ? 'Expand'
                         : 'Collapse';
                 }